+++ R/RcppExports.R
@@ -1,15 +1,15 @@
 # This file was generated by Rcpp::compileAttributes
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-opt_montecarlocpp <- function(Xc, current, candidateidx, crit, iterations, repeated) {
-    .Call('OptimalDesign_opt_montecarlocpp', PACKAGE = 'OptimalDesign', Xc, current, candidateidx, crit, iterations, repeated)
+initDesign <- function(X, design_size, repeated) {
+    .Call('OptimalDesign_initDesign', PACKAGE = 'OptimalDesign', X, design_size, repeated)
 }
 
 opt_geneticrealcpp <- function(parents, n, iterations, pidx) {
     .Call('OptimalDesign_opt_geneticrealcpp', PACKAGE = 'OptimalDesign', parents, n, iterations, pidx)
 }
 
-initDesign <- function(X, design_size, repeated) {
-    .Call('OptimalDesign_initDesign', PACKAGE = 'OptimalDesign', X, design_size, repeated)
+opt_montecarlocpp <- function(Xc, current, candidateidx, crit, iterations, repeated) {
+    .Call('OptimalDesign_opt_montecarlocpp', PACKAGE = 'OptimalDesign', Xc, current, candidateidx, crit, iterations, repeated)
 }
